+++ resources/views/reddit.blade.php
@@ -2,7 +2,5 @@
     <blockquote class="reddit-embed-bq" data-embed-height="500">
         <a href="{!! $url ?? '#' !!}"></a>
     </blockquote>
-    <aside class="shortcode_nocookie">
-        {!! config('shortcode-plus.nocookie.text') !!} - Reddit.com
-    </aside>
+    @include('shortcode-plus::components.paywall')
 </div>